% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainworkupr-package.R
\docType{package}
\name{brainworkupr-package}
\alias{brainworkupr}
\alias{brainworkupr-package}
\title{brainworkupr: BrainWorkUp R Neuropsychological Report Writing Package}
\description{
This package stores the functions and normative datasets to use for neuropsychological assessments at the USC Keck School of Medicine Department of Psychiatry and Behavioral Sciences.
}
\author{
\strong{Maintainer}: Joey Trampush \email{trampush@usc.edu} (\href{https://orcid.org/0000-0003-0830-508X}{ORCID})

Other contributors:
\itemize{
  \item USC [copyright holder]
}

}
\keyword{internal}
