% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drill.R
\name{drill}
\alias{drill}
\title{Drilldown on neuropsych domains.}
\usage{
drill(patient = NULL, data)
}
\arguments{
\item{patient}{Patient name}

\item{data}{Dataset to use}
}
\value{
A drilldown plot
}
\description{
This function uses the R Highcharter package and drilldown function to "drill down" on neuropsychological domains and test scores.
}
\examples{
datafile <- system.file("neurocog.csv", package = "brainworkupr")
neurocog <- read.csv(datafile, encoding = "UTF-8")
drill(patient = "Biggie", data = neurocog)
}
