% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_df.R
\name{make_df}
\alias{make_df}
\title{Make data frame}
\usage{
make_df(params = NULL)
}
\value{
A data frame
}
\description{
Make data frame
}
\examples{
df <- make_df(
  params = list(
    test = "tmt",
    test_name = "Trail Making Test",
    scale = "Trails B",
    raw_score = 55,
    score = 50,
    score_type = "t_score",
    domain = "Attention/Executive",
    subdomain = "Processing Speed",
    narrow = "Set-Shifting",
    pass = "Planning",
    verbal = "Nonverbal",
    timed = "Timed",
    mean = 50,
    stdev = 10,
    reliability = .95,
    test_type = "npsych_test"
  )
)

}
